separated mappers and started adding tests
